+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.4.0beta] - 2016-10-xx
+### Added
+- `bower.json` Optional
+
 ### Changed
 - Moved bower to npm
 - Updated dependencies